Code refactore
Moved input configuration processing related code inside new package
`inputprocessor`.
